+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased](https://github.com/model-bakers/model_bakery/tree/main)
 
 ### Added
+- Add support to `auto_now` and `auto_now_add` fields.
 
 ### Changed
 - Remove unnecessary casting to string methods random_gen.gen_slug and random_gen.gen_string